requirements.txt
streamlit
langchain-openai
langchain-pinecone
pinecone-client
openai
python-dotenv